requirements.txt
dj-database-url==0.5.0
Django==3.1.1
django-crispy-forms==1.9.2
django-heroku==0.3.1
django-markdownify==0.8.1
django-social-share==2.0.0
django-taggit==1.3.0
easy-thumbnails==2.7
gunicorn==20.0.4
Markdown==3.3
Pillow==7.2.0
sqlparse==0.3.1
whitenoise==5.2.0
